+++ plugins/wikitext/tests/misc.test.js
@@ -1,8 +1,8 @@
 var test = require('tape')
 var wtf = require('./_lib')
 
-test('basic-html', t => {
-  let arr = ['that cat is [[a]] cool dude', 'cool [[stuff]] **bold** too']
+test('basic-wikitext', t => {
+  let arr = ['that cat is [[a]] cool dude', `one [[two|2]] three`, 'cool [[stuff]] **bold** too']
   arr.forEach(str => {
     let doc = wtf(str)
     t.equal(doc.wikitext(), str, str)