+++ tests/templates-inline.test.js
@@ -162,7 +162,7 @@ test('flags', function(t) {
   var doc = wtf(str)
   t.equal(doc.links().length, 3, 'found 3 link')
   t.equal(doc.links(1).text(), 'DEU', 'link text')
-  t.equal(doc.links(1).page(), 'Germany', 'link page')
+  t.equal(doc.links(1).page(), 'germany', 'link page')
   t.equal(doc.text(), 'one 🇺🇸 USA, two 🇩🇪 DEU, three 🇨🇦 CAN.', 'made emoji flags')
   t.end()
 })